New way to get faster.
With this method, we are going faster and we consume less memory
than before (I would say that we save 73 bytes per checksums at
least). I added a new field to the data_t * structure that is
a boolean which is TRUE if the checksum and it's associated data
is in the cache and FALSE if it is not in the cache.