Trying to limit sizes of asynchronous queues
There is a performance problem with queues and the store queue is
not so receiving data as quickly as others queues does. Trying to
mimit the number of elements in queues was a test to see if this
was applicable. But it seems that sometimes a queue never receives
data that should have been sent (may be a cache effect) and thus
everyone is waiting for something that will never happen !