The error was that smeta variable was used in two thread simultaneously.
The problem was that it was freed in one thread and used in another one!
That was leading to some mistakes such as the pointer existed when
doing the test but did not anymore when accessing the datas ->
segfaults.
To avoid this I put the code using the smeta variable before the "call"
to the thread and the thread MUST free the variable when done (ie this
is the backend's function job).