Uses different names to save files
This is done to avoid overwriting an existing file. This should not happen.
A number is added at the end of the date when saving all versions of a file
of the same date. A number is added in front of the file if restoring only
one file that already exists.